Create Task 2 use info long_rivers.py
In this task for week 2, we were given a dictionary list of rivers. We had to find the total length of all the rivers, print out the name of all the rivers, calculate the mean of the rivers, and print out the name of the rivers that started with the letter "M".